add models file
